+++ src/components/Visualizations/GraphPanes/index.ts
@@ -1,4 +1,5 @@
 export * from './HIndexGraphPane';
 export * from './AuthorNetworkGraphPane';
+export * from './MetricsPane';
 export * from './PaperNetworkGraphPane';
 export * from './YearsGraphPane';
