+++ src/components/Visualizations/Graphs/AuthorNetworkGraph.tsx
@@ -1,4 +1,4 @@
-import { useD3 } from '../useD3';
+import { useD3 } from './useD3';
 import * as d3 from 'd3';
 import { ReactElement, useCallback, useEffect, useMemo, useState } from 'react';
 import { BaseType, D3ZoomEvent, HierarchyRectangularNode, Selection } from 'd3';
