+++ src/components/Visualizations/Panes/NetworkDetails/PaperNetworkDetailsPane.tsx
@@ -16,8 +16,8 @@ import {
   Tbody,
 } from '@chakra-ui/react';
 import { SimpleLink } from '@components/SimpleLink';
+import { ILineGraph } from '@components/Visualizations/types';
 import { ReactElement, useEffect, useState } from 'react';
-import { ILineGraph } from '../../types';
 import { NodeDetailPane } from './NodeDetailsPane';
 import { SummaryPane } from './SummaryPane';
