+++ src/components/Visualizations/Widgets/NotEnoughData.tsx
@@ -1,5 +1,5 @@
 import { Box, Text } from '@chakra-ui/react';
-import { SimpleLink } from '@components/SimpleLink';
+import { SimpleLink } from '@components';
 import { ReactElement } from 'react';
 
 export const NotEnoughData = (): ReactElement => {
