+++ src/components/Visualizations/index.tsx
@@ -1,7 +1,7 @@
 export * from './Containers';
 export * from './GraphPanes';
 export * from './Graphs';
-export * from './Metrics';
 export * from './Panes';
 export * from './Tables';
 export * from './VisualizationTabs';
+export * from './types';
